+++ CHANGELOG.md
@@ -62,4 +62,4 @@
 
 * Add the method <code>Future<PrinterStatus> imprimirCodigoDeBarraImg(String mensagem, int width, int height, BarcodeTypes barcodeType)</code>.
 * Remove the method <code>Future<PrinterStatus> fimImpressao()</code> and call it after every printer command.
-* Updated example.
+* Updated README.md and example.